requirements.txt
streamlit 
langchain 
openai 
faiss-cpu 
python-dotenv 
unstructured
langchain_community
tiktoken